More resilient version for webOS 3.x (v0.1.4)
Version 0.1.3 (and previous versions) were not working accross more
than two reboots, at least under webOS 3.4. This is because we get
relaunched only on the reboot immediately following a successful
addDevice call, so this version now ensures that we remove/re-add
ourselves from the input app list every time, which guarantees that
it works for all next reboots.

The relaunch of the previous input app is now done as soon as possible,
because this is what gives back control to the user, while we can continue
our setup in the background.